Test: if statement

+++ test/travis_ci_scripts/upload_nebula_script.sh
@@ -9,11 +9,17 @@ git clone --quiet --branch=test-branch-1  https://$USER_NAME:$GITHUB_API_KEY@git
 cd master
 pyinstaller -F nebulactl.py
 
-#add, commit and push files
-git add -f .
-git remote rm origin
-git remote add origin https://$USER_NAME:$GITHUB_API_KEY@github.com/$USER_NAME/nebula-cmd.git > /dev/null 2>&1
-git add -f .
-git commit -m "Travis build $TRAVIS_BUILD_NUMBER pushed - nebulactl.py run and pushed"
-git push --quiet --set-upstream origin test-branch-1
-echo -e "Done"
+MESSAGE=$(git log -1 HEAD --pretty=format:%s)
+
+if [[ "$MESSAGE" == "Travis build" ]]; then
+    echo "already pushed run grunt and exit"
+else
+    #add, commit and push files
+    git add -f .
+    git remote rm origin
+    git remote add origin https://$USER_NAME:$GITHUB_API_KEY@github.com/$USER_NAME/nebula-cmd.git > /dev/null 2>&1
+    git add -f .
+    git commit -m "Travis build $TRAVIS_BUILD_NUMBER pushed - nebulactl.py run and pushed"
+    git push --quiet --set-upstream origin test-branch-1
+    echo -e "Done"
+fi
